% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FizzBuzz.R
\name{FizzBuzz}
\alias{FizzBuzz}
\title{Fizz Buzz}
\usage{
FizzBuzz(n)
}
\arguments{
\item{n}{An integer that must be greater than 1 and cannot be infinity.}
}
\value{
A character vector of length \code{n} containing integer values and characters.
}
\description{
Fizz Buzz is a childrens game where a group of people take turns counting,
if the number you are about to say is divisible by 3 you shout Fizz and if it
is divisible by 5 you shout Buzz. If the number is divisible by both 3 and
15 i.e. divisible by 15 you shout FizzBuzz.
}
\details{
This function returns a list of length n containing numbers ranging from
1 to n where if the number is divisible by 3,5, or 15 it replaces it with Fizz,
Buzz, and FizzBuzz respectivly
}
\examples{
FizzBuzz(12)
FizzBuzz(30)
}
